requirements.txt
langchain
langchain-openai
langchain-community
duckduckgo-search
python-pptx
requests
python-dotenv